Move Auth to new package
