+++ src/main/java/uk/gov/mca/beacons/api/auth/gateway/AuthGateway.java
@@ -1,8 +1,8 @@
-package uk.gov.mca.beacons.api.gateways;
+package uk.gov.mca.beacons.api.auth.gateway;
 
 import java.util.List;
+import uk.gov.mca.beacons.api.auth.gateway.AuthGatewayImpl.SupportedPermissions;
 import uk.gov.mca.beacons.api.domain.User;
-import uk.gov.mca.beacons.api.gateways.AuthGatewayImpl.SupportedPermissions;
 
 public interface AuthGateway {
   User getUser();
