+++ src/test/java/uk/gov/mca/beacons/api/hateoas/HateoasLinkManagerTest.java
@@ -3,7 +3,7 @@ package uk.gov.mca.beacons.api.hateoas;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.mockito.BDDMockito.given;
-import static uk.gov.mca.beacons.api.gateways.AuthGatewayImpl.SupportedPermissions;
+import static uk.gov.mca.beacons.api.auth.gateway.AuthGatewayImpl.SupportedPermissions;
 
 import java.util.ArrayList;
 import java.util.UUID;
@@ -11,8 +11,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import uk.gov.mca.beacons.api.auth.application.BeaconRolesService;
 import uk.gov.mca.beacons.api.jpa.entities.Beacon;
-import uk.gov.mca.beacons.api.services.BeaconRolesService;
 
 @ExtendWith(MockitoExtension.class)
 class HateoasLinkManagerTest {
